+++ bin/sam-reference-read-counts.py
@@ -2,14 +2,20 @@
 
 from __future__ import print_function, division
 
+import sys
 import argparse
 from collections import defaultdict
 
-from dark.sam import samfile
+from dark.sam import samfile, SAMFilter
 from dark.utils import pct
 
 
 def referenceInfo():
+    """
+    Hold information about the reads that match a reference.
+
+    @return: A C{dict}, as below.
+    """
     return {
         'duplicate': set(),
         'primary': set(),
@@ -22,10 +28,23 @@ def referenceInfo():
 
 
 def main(args):
+    """
+    Print SAM/BAM file reference read counts.
+
+    @param args: An argparse namespace with information about parsed
+        command-line options.
+    """
+    if args.topReferenceIdsFile and args.sortBy != 'count':
+        print('--topReferenceIdsFile only makes sense when using --sortBy '
+              'count', file=sys.stderr)
+        sys.exit(1)
+
     referenceReads = defaultdict(referenceInfo)
     mapped = set()
     unmapped = set()
     readIds = set()
+
+    referenceLengths = SAMFilter(args.samFile).referenceLengths()
 
     with samfile(args.samFile) as fp:
         for read in fp.fetch():
@@ -55,33 +74,42 @@ def main(args):
 
     totalReads = len(readIds)
 
-    print('Found a total of %d read%s (%d mapped, %d unmapped).' %
+    print('Found a total of %d read%s (%d mapped, %d unmapped) mapping '
+          'against %d of %d reference%s.' %
           (totalReads, '' if totalReads == 1 else 's',
-           len(mapped), len(unmapped)))
+           len(mapped), len(unmapped),
+           len(referenceReads), len(referenceLengths),
+           '' if len(referenceLengths) == 1 else 's'))
 
-    def key(referenceId):
-        return len(referenceReads[referenceId]['readIds'])
+    if args.sortBy == 'count':
 
-    sortedReferenceReads = sorted(referenceReads, key=key, reverse=True)
-    topReference = sortedReferenceReads[0]
+        def key(referenceId):
+            return len(referenceReads[referenceId]['readIds'])
 
-    if not args.sort:
-        # Just sort the references by name, not by number of reads.
+        sortedReferenceReads = sorted(referenceReads, key=key, reverse=True)
+        topReference = sortedReferenceReads[0]
+    else:
+        # Sort the references by name
         sortedReferenceReads = sorted(referenceReads)
 
     cumulativeReadIds = set()
 
-    for referenceId in sortedReferenceReads:
+    for count, referenceId in enumerate(sortedReferenceReads, start=1):
         stats = referenceReads[referenceId]
         readCount = len(stats['readIds'])
+        if readCount == 0 and args.excludeZeroes:
+            continue
         newReadCount = len(stats['readIds'] - cumulativeReadIds)
+        if newReadCount == 0 and args.excludeIfNoAdditional:
+            continue
         cumulativeReadIds.update(stats['readIds'])
-        print('\n%s:\n'
+        print('\nReference %d: %s (%d nt):\n'
               '  Overall reads mapped to the reference: %s\n'
               '  Non-duplicates: %s, Duplicates: %s, QC fails: %s\n'
               '  Primary: %s, Secondary: %s, Supplementary: %s\n'
-              '  Reads not matching any reference above: %s' %
-              (referenceId,
+              '  Reads not matching any reference above: %s\n'
+              '  Previously unmatched reads for this reference: %s' %
+              (count, referenceId, referenceLengths[referenceId],
                pct(readCount, totalReads),
                pct(len(stats['nonDuplicate']), readCount),
                pct(len(stats['duplicate']), readCount),
@@ -89,7 +117,8 @@ def main(args):
                pct(len(stats['primary']), readCount),
                pct(len(stats['secondary']), readCount),
                pct(len(stats['supplementary']), readCount),
-               pct(newReadCount, totalReads)))
+               pct(newReadCount, totalReads),
+               pct(newReadCount, readCount)))
 
     # Write out the (sorted) read ids of the reference with the most reads.
     if args.topReferenceIdsFile:
@@ -109,13 +138,27 @@ if __name__ == '__main__':
         help='The name of a SAM/BAM alignment file.')
 
     parser.add_argument(
-        '--sort', action='store_true',
-        help=('Sort the output by decreasing read count (i.e., the reference '
-              'with the highest number of matching reads is printed first).'))
+        '--sortBy', choices=('count', 'name'), default='count',
+        help=('Use "count" to sort the output by decreasing read count (i.e., '
+              'the reference with the highest number of matching reads is '
+              'printed first), or "name" to sort by reference name.'))
+
+    parser.add_argument(
+        '--excludeZeroes', action='store_true',
+        help='Do not print references unmatched by any read.')
+
+    parser.add_argument(
+        '--excludeIfNoAdditional', action='store_true',
+        help=('Do not print references that are mapped by no additional reads '
+              '(i.e., reads other than those matching already-printed '
+              'references). This is useful when --sortBy count is used, '
+              'making it possible to not print references only matched by '
+              'reads matching references that have already been printed.'))
 
     parser.add_argument(
         '--topReferenceIdsFile', metavar='FILE',
         help=('The file to write the (sorted) ids of the reads for the '
-              'reference with the highest number of matching reads.'))
+              'reference with the highest number of matching reads. Only '
+              'valid when --sortBy count is used.'))
 
     main(parser.parse_args())